Workaroud MacOS/SMB bug
When an outdated tag is found on an SMB filesystem on a Mac, it seems the call
to `fsetxattr` does not update the xattr but removes it instead. So it takes
two runs of `cshatag` to update the attribute. To work around this issue, we
remove the xattr explicitely before setting it again.